Move debug adapter to application package
